requirements.txt
streamlit==1.38.0
langchain==0.3.0
langchain-community==0.3.0
transformers==4.44.2
torch==2.4.1
faiss-cpu==1.8.0
pandas==2.2.2
numpy==1.26.4
packaging==23.2